added some tests
